Added Device API
